chore: refactor azure/gcp code
Extract common code under cloud/utils.h
