+++ util/cloud/azure/CMakeLists.txt
@@ -1,2 +1,2 @@
-add_library(azure_lib azure.cc)
+add_library(azure_lib azure.cc storage.cc)
 cxx_link(azure_lib http_client_lib strings_lib TRDP::pugixml)
